requirements.txt
keras
flask
requests
numpy
pillow
# tensorflow
# tensorflow-serving-api
